chatbot/pyproject.toml
[tool.poetry]
name = "chatbot"
version = "0.0.1"
description = "A simple example to build a chatbot using Restack"
authors = [
    "Restack Team <service@restack.io>",
]
readme = "README.md"
packages = [{include = "src"}]

[tool.poetry.dependencies]
python = ">=3.10,<4.0"
openai = "^1.57.2"
python-dotenv = "1.0.1"
restack-ai = "^0.0.54"
watchfiles = "^1.0.0"
pydantic = "^2.10.4"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.scripts]
dev = "src.services:watch_services"
services = "src.services:run_services"
schedule = "schedule_workflow:run_schedule_workflow"
